requirements.txt
Flask
flask-cors
pdfplumber
python-docx
gunicorn
